feat: 重构配置系统并优化训练数据生成
refactor(config): 重写配置类以支持JSONC格式配置文件
feat(config): 新增嵌套配置值获取方法
fix(config): 修复数据库路径配置项名称不一致问题

refactor(generate_training_data): 优化消息处理流程
feat(generate_training_data): 添加中断信号处理和多线程支持
fix(generate_training_data): 修复图片内容检测逻辑

docs: 新增CLI设计文档
chore: 删除旧的generate_training_data_llm.py文件
