+++ seeting.jsonc
@@ -1,0 +1,70 @@
+{
+    "repoid": "Qing-Agent",
+    "branch": "main",
+    "version": "0.1.2",
+    "model_args": {
+        "model_path": "./model/Qwen3-8B-Base",
+        "model_repo": "Qwen/Qwen-3-8B-Base", //如果不存在model_path则自动下载model_repo
+        "model_output_path": "./model_output",
+        "template": "qwen",
+        "finetuning_type": "qlora",
+        "trust_remote_code": true,
+        "download_source": "modelscope"
+    },
+    "logger_args": {
+        "log_level": "INFO",
+        "language": "zhcn"
+    },
+    "data_args": {
+        "qq_agrs":{
+            "qq_db_path": "./data/qq.db",
+            "qq_number_ai": "1234567890"
+        },
+        "telegram_args": {
+            "telegram_chat_id": "user1234567890"
+        },
+        "include_type": [
+            "text"
+        ],
+        "blocked_words": [ // 禁用词
+            "密码",
+            "账号",
+            "手机号"
+        ],
+        "single_combine_time_window": 2, // 单条消息合并时间窗口，单位：分钟
+        "qa_match_time_window": 5, // 组成qa时间窗口（分钟）,
+        "combine_msg_max_length": 2048, // 组合后消息最大长度
+        "messages_max_length": 2048, // messages最长字符数量
+        "clean_set_args": {
+            "clean_method": "raw", //或者使用raw,纯算法清理,速度更快
+            "openai_api": { //可以使用网站API或者vLLM或者LMStudio
+                "api_base": "https://api.openai.com/v1",
+                "api_key": "sk-1234567890abcdef1234567890abcdef",
+                "model_name": "xxx",
+                "clean_batch_size": 10,
+                "clean_workers": 4,
+                "timeout": 120
+            }
+        },
+        "train_sft_args": {
+            "use_qlora" :true,
+            "data_path": "./data/sft.jsonl",
+            "lora_r":16,
+            "lora_alpha":32,
+            "lora_dropout":0.1,
+            "lora_target_modules": [
+                "q_proj",
+                "v_proj"
+            ],
+            "tper_device_train_batch_size": 1,
+            "per_device_eval_batch_size": 1,
+            "gradient_accumulation_steps": 16,
+            "max_steps": -1,
+            "learning_rate": 2e-4,
+            "fp16": true,
+            "logging_steps": 10,
+            "save_steps": 100,
+            "load_precision": "int8"
+        }
+    }
+}
